V3: Angular 8 / Java 11
* Drop AngularJS and Thymleaf
* Use Clarity Design instead of Boostrap
* Use Spring-boot 2.2 instead of 1.5

+++ trading-common/src/main/java/fr/ymanvieu/trading/portofolio/OrderException.java
@@ -26,11 +26,11 @@ public class OrderException extends BusinessException {
 		super(key, args);
 	}
 
-	public static OrderException NOT_ENOUGH_FUND(String code, float quantity, String curCode, float available, float needed) {
+	public static OrderException NOT_ENOUGH_FUND(String code, double quantity, String curCode, double available, double needed) {
 		return new OrderException("order.error.not_enough_fund", quantity, code, curCode, available, needed);
 	}
 
-	public static OrderException NOT_ENOUGH_OWNED(String code, float owned, float toSell) {
+	public static OrderException NOT_ENOUGH_OWNED(String code, double owned, double toSell) {
 		return new OrderException("order.error.not_enough_owned", code, owned, toSell);
 	}
